requirements.txt
click
flask
flask-socketio
flask-cors
psycopg2-binary
sqlalchemy